@@ -2,7 +2,6 @@ package check
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/cashapp/spirit/pkg/utils"
 	"github.com/siddontang/loggers"
@@ -15,6 +14,6 @@ func init() {
 // illegalClauseCheck checks for the presence of specific, unsupported
 // clauses in the ALTER statement, such as ALGORITHM= and LOCK=.
 func illegalClauseCheck(ctx context.Context, r Resources, logger loggers.Advanced) error {
-	sql := fmt.Sprintf("ALTER TABLE x.x %s", r.Alter)
+	sql := "ALTER TABLE x.x " + r.Alter
 	return utils.AlterContainsUnsupportedClause(sql)
 }
