@@ -25,7 +25,7 @@ import (
 
 func TestReserveBuffer(t *testing.T) {
 	res0 := reserveBuffer(nil, 0)
-	require.Len(t, res0, 0)
+	require.Empty(t, res0)
 
 	res1 := reserveBuffer(res0, 3)
 	require.Len(t, res1, 3)
@@ -96,8 +96,6 @@ func TestEscapeBackslash(t *testing.T) {
 		},
 	}
 	for _, v := range tests {
-		// copy iterator variable into a new variable, see issue #27779
-		v := v
 		t.Run(v.name, func(t *testing.T) {
 			require.Equal(t, v.output, escapeBytesBackslash(nil, v.input))
 			require.Equal(t, v.output, escapeStringBackslash(nil, string(v.input)))
@@ -402,8 +400,6 @@ func TestEscapeSQL(t *testing.T) {
 		},
 	}
 	for _, v := range tests {
-		// copy iterator variable into a new variable, see issue #27779
-		v := v
 		t.Run(v.name, func(t *testing.T) {
 			r3 := new(strings.Builder)
 			r1, e1 := escapeSQL(v.input, v.params...)
